move GPGKeypairManagerinto asserts, address easy error message feedback
